Completed Day 3 Part 2
Went down the wrong path with Part 1 and didn't have time to refactor to make Part 2 easier.  Really needs to be refactored as code is far too complicated for the task.

Tried to do it with a single pass of the input file which I think in hindsight is probably an unnecessary complication.