Completed Day 5 Part 2
Not particularly performant so definitely missing a trick somewhere along the way.  Must be traversing too many paths that I don't need to so expect there is a 'fail early' exit point I haven't picked up on, or an alternate way of navigating the maps using some sort of indexing strategy?